# Kyber Aarch64: assembly implementations of functions