… wolfcrypt/src/port/arm/armv8-kyber-asm.S
… wolfcrypt/src/port/arm/armv8-kyber-asm_c.c
… wolfcrypt/src/wc_kyber_poly.c