# Description

Aarch64 assembly implementation of Kyber functions.
SHA-3 assembly implementations when not hardware crypto.

# Testing

Tested on M1 Mac.
Inline assembly tested as well.

# Checklist

 - [ ] added tests
 - [ ] updated/added doxygen
 - [ ] updated appropriate READMEs
 - [ ] Updated manual and documentation