@dgarske 2024-09-26
Tested on Xilinx UltraScale+ ZCU102:

Master:
```
# ./benchmark -kyber -sha3
------------------------------------------------------------------------------
 wolfSSL version 5.7.2
------------------------------------------------------------------------------
Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
        Single Precision: ecc 256 384 521 rsa 2048 3072 4096 asm sp_arm64.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
SHA3-224                    65 MiB took 1.026 seconds,   63.345 MiB/s
SHA3-256                    65 MiB took 1.080 seconds,   60.209 MiB/s
SHA3-384                    50 MiB took 1.057 seconds,   47.304 MiB/s
SHA3-512                    35 MiB took 1.039 seconds,   33.692 MiB/s
KYBER512    128  key gen      8900 ops took 1.005 sec, avg 0.113 ms, 8859.861 ops/sec
KYBER512    128    encap      7600 ops took 1.012 sec, avg 0.133 ms, 7508.676 ops/sec
KYBER512    128    decap      5400 ops took 1.011 sec, avg 0.187 ms, 5338.993 ops/sec
KYBER768    192  key gen      5200 ops took 1.005 sec, avg 0.193 ms, 5173.026 ops/sec
KYBER768    192    encap      4600 ops took 1.013 sec, avg 0.220 ms, 4539.985 ops/sec
KYBER768    192    decap      3400 ops took 1.004 sec, avg 0.295 ms, 3385.359 ops/sec
KYBER1024   256  key gen      3300 ops took 1.027 sec, avg 0.311 ms, 3213.490 ops/sec
KYBER1024   256    encap      3000 ops took 1.030 sec, avg 0.343 ms, 2911.320 ops/sec
KYBER1024   256    decap      2300 ops took 1.007 sec, avg 0.438 ms, 2284.343 ops/sec
Benchmark complete
```

PR 7998:
```
./benchmark -kyber -sha3
------------------------------------------------------------------------------
 wolfSSL version 5.7.2
------------------------------------------------------------------------------
Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
        Single Precision: ecc 256 384 521 rsa 2048 3072 4096 asm sp_arm64.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
SHA3-224                    80 MiB took 1.056 seconds,   75.769 MiB/s
SHA3-256                    75 MiB took 1.039 seconds,   72.165 MiB/s
SHA3-384                    60 MiB took 1.054 seconds,   56.931 MiB/s
SHA3-512                    45 MiB took 1.102 seconds,   40.837 MiB/s
KYBER512    128  key gen     19900 ops took 1.002 sec, avg 0.050 ms, 19863.217 ops/sec
KYBER512    128    encap     17800 ops took 1.001 sec, avg 0.056 ms, 17790.202 ops/sec
KYBER512    128    decap     12500 ops took 1.005 sec, avg 0.080 ms, 12440.752 ops/sec
KYBER768    192  key gen     12700 ops took 1.003 sec, avg 0.079 ms, 12662.974 ops/sec
KYBER768    192    encap     11400 ops took 1.005 sec, avg 0.088 ms, 11343.019 ops/sec
KYBER768    192    decap      8300 ops took 1.009 sec, avg 0.122 ms, 8228.787 ops/sec
KYBER1024   256  key gen      7600 ops took 1.013 sec, avg 0.133 ms, 7503.618 ops/sec
KYBER1024   256    encap      7000 ops took 1.012 sec, avg 0.145 ms, 6918.733 ops/sec
KYBER1024   256    decap      5400 ops took 1.007 sec, avg 0.186 ms, 5364.265 ops/sec
Benchmark complete
```

Cross compiled using: `./configure --host=aarch64 CC="aarch64-linux-gnu-gcc" AR="aarch64-linux-gnu-ar" RANLIB="aarch64-linux-gnu-ranlib" --enable-sp=yes,asm --enable-keygen --enable-armasm --enable-experimental --enable-kyber --enable-dilithium --enable-keygen --enable-lms --enable-xmss --enable-curve25519 --enable-ed25519 --enable-curve448 --enable-ed448 --disable-shared --enable-static --disable-dh --disable-filesystem && make`